requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy
greenlet
asyncpg
google-genai
pydantic-settings